update: user routing on login, register and forgot password
